fix:migrate leave absent

+++ database/migrations/2024_09_17_020422_add_type_absent_to_absent_requests_table.php
@@ -12,9 +12,15 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('absent_requests', function (Blueprint $table) {
-            $table->string('type_absent')->nullable();
-            $table->foreignId('hrd_id')->nullable()->constrained('employees')->nullOnDelete()->cascadeOnUpdate();
-            $table->timestamp('hrd_approved_at')->nullable();
+            if (Schema::hasColumn('absent_requests', 'type_absent')) {
+                $table->string('type_absent')->nullable();
+            }
+            if (Schema::hasColumn('absent_requests', 'hrd_id')) {
+                $table->foreignId('hrd_id')->nullable()->constrained('employees')->nullOnDelete()->cascadeOnUpdate();
+            }
+            if (Schema::hasColumn('absent_requests', 'hrd_approved_at')) {
+                $table->timestamp('hrd_approved_at')->nullable();
+            }
         });
     }
 
@@ -24,7 +30,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('absent_requests', function (Blueprint $table) {
-            $table->dropColumn('type_absent');
+            if (Schema::hasColumn('absent_requests', 'type_absent')) {
+                $table->dropColumn('type_absent');
+            }
             if (Schema::hasColumn('absent_requests', 'hrd_id')) {
                 $table->dropColumn('hrd_id');
             }
